backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
httpx>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
# Will be needed in Phase 3
# langgraph>=0.0.10
# openai>=1.1.0 